[REFACTOR] #24 거래 내역 금액 및 계좌 잔액 타입 변경

+++ backend/src/main/java/com/ssafy/moyeobang/settle/application/domain/order/MemberOrderHistory.java
@@ -10,11 +10,11 @@ public class MemberOrderHistory {
 
     private final Long id;
 
-    private final Integer amount;
+    private final long amount;
 
     private MappingInfo mappingInfo;
 
-    public static MemberOrderHistory of(Long id, Integer amount, MappingInfo mappingInfo) {
+    public static MemberOrderHistory of(Long id, Long amount, MappingInfo mappingInfo) {
 
         return new MemberOrderHistory(id, amount, mappingInfo);
     }
